Updated to follow same plug structure as Iwan. Changed labels to be experiment and phase tags.
