+++ Makefile
@@ -20,7 +20,6 @@ lint:
 	fgt golint ./pkg/...
 	fgt golint ./experiments/...
 	fgt golint ./cmds/...
-	# fgt golint ./plugin/...
 	fgt golint ./misc/...
 
 unit_test:
@@ -29,11 +28,16 @@ unit_test:
 	go test $(TEST_OPT) ./cmds/...
 	go test $(TEST_OPT) ./misc/...
 
-integration_test:
+plugins:
+	(cd build; go build ../misc/snap-plugin-collector-session-test)
+	(cd build; go build ../misc/snap-plugin-processor-session-tagging)
+	(cd build; go build ../misc/snap-plugin-publisher-session-test)
+	(cd misc/snap-plugin-collector-mutilate; go build)
+
+integration_test: plugins
 	go test $(TEST_OPT) -tags=integration ./pkg/...
 	go test $(TEST_OPT) -tags=integration ./experiments/...
 	go test $(TEST_OPT) -tags=integration ./cmds/...
-	(cd misc/snap-plugin-collector-mutilate; go build; cd ../..)
 #   TODO(niklas): Fix race (https://intelsdi.atlassian.net/browse/SCE-316)
 #	go test $(TEST_OPT) -tags=integration ./misc/...
 	go test -tags=integration ./misc/...
